Log time bug fixed; decreased sample lag
1) The log time is no longer copied from the most recent sample, but
instead calculated at the time of the log. PC and EyeTribe clock
asynchrony is calculated at every new incoming sample, so it shouldn't
be too inaccurate.

2) The most recent sample (returned by the `sample` and `pupil_size`
methods) is no longer set in the sample processing Thread, but directly
as it comes in. This should fix (part of) the lag that some users were
experiencing.